Move CompletableEmitterGmxTaskHandler to util package
